TPZSolver and TPZMatrixSolver are now declared in different files
