+++ Mesh/pzcompel.cpp
@@ -19,7 +19,7 @@
 #include "pzmatrix.h"
 #include "pzfmatrix.h"
 #include "pzblock.h"
-#include "pzsolve.h"
+#include "TPZMatrixSolver.h"
 
 #include "pzmanvector.h"
 #include "pzstack.h"
