+++ Solvers/TPZSolver.cpp
@@ -1,0 +1,6 @@
+#include "TPZSolver.h"
+#include "Hash/TPZHash.h"
+
+int TPZSolver::ClassId() const{
+    return Hash("TPZSolver");
+}
